UI i filament

+++ resources/js/utils/date.js
@@ -1,16 +1,24 @@
+/**
+ * Transforms date to a string with the format: "Wed, Jan 20, 2021"
+ * @param {*} date 
+ * @returns 
+ */
 const dayFormatted = (date) => {
   let day = date;
   if (typeof date === "string") {
     day = new Date(date);
   }
   return Intl.DateTimeFormat("ca-es", {
-    weekday: "long",
+    weekday: "short",
     month: "short",
     day: "numeric",
     year: "numeric",
   }).format(day);
 };
 
+/**
+ * Transforms date to a string with the format: "2021-01-20"
+ */
 const ymd = (date) => {
   if (!date) {
     return "";
